move the report functions to a separate module
